requirements.txt
Pillow>=9.0.0
numpy>=1.20.0
tkinterdnd2>=0.4.0
